moved player ground detection to player.rs seperated player movement state updates refined player hitbox
